20250813 로그 조회 sql 수정

+++ frontend/src/views/auth/Login.vue
@@ -57,12 +57,14 @@ const togglePassword = () => {
 // 화면진입 시
 onMounted(async () => {
 
-  await Api.post(ApiUrls.LOGOUT, {}, true);
-
-  setTimeout(()=>{
-    userStore().delUserInfo();
-    sessionStorage.clear();
-  }, 1000);
+  const isLoggedIn = userStore().isLoggedIn;
+  if (isLoggedIn) {
+    await Api.post(ApiUrls.LOGOUT, {}, true);
+    setTimeout(()=>{
+      userStore().delUserInfo();
+      sessionStorage.clear();
+    }, 1000);
+  }
 
   state.isVisible = false;
   state.isProcessing = false;
